make a clean file that just is for parsing the command line
