+++ .gitignore
@@ -1,0 +1,1 @@
+community/seascape-shader/seascape-shader
