+++ community/seascape-shader/seascape.go
@@ -2,27 +2,12 @@ package main
 
 import (
 	"time"
-	"log"
 
 	"github.com/go-gl/mathgl/mgl32"
 
 	"github.com/faiface/pixel"
 	"github.com/faiface/pixel/pixelgl"
 	"golang.org/x/image/colornames"
-)
-
-import "flag"
-import "os"
-
-var (
-	version       string
-	race          bool
-	debug         = os.Getenv("BUILDDEBUG") != ""
-	filename      string
-	width         int
-	height        int
-	timeout       = "120s"
-	uDrift        float32
 )
 
 func run() {
@@ -83,25 +68,6 @@ func run() {
 
 		win.Update()
 	}
-
-}
-
-func parseFlags() {
-	flag.StringVar  (&version,       "version",       "v0.1",                     "Set compiled in version string")
-	flag.StringVar  (&filename,      "filename",      "shaders/seascape.glsl",    "path to GLSL file")
-	flag.IntVar     (&width,         "width",         1024,                       "Width of the OpenGL Window")
-	flag.IntVar     (&height,        "height",        768,                        "Height of the OpenGL Window")
-	var tmp float64
-	flag.Float64Var (&tmp,           "drift",         0.01,                       "Speed of the gradual camera drift")
-	flag.BoolVar    (&race,          "race",          race,                       "Use race detector")
-
-	// this parses the arguements
-	flag.Parse()
-
-	uDrift = float32(tmp)
-	log.Println("width=",width)
-	log.Println("height=",height)
-	log.Println("uDrift=",uDrift)
 
 }
